Security, Privacy and Explainability/requirements.txt
bcrypt
pyjwt
flask
flask-limiter
mysql-connector-python
python-dotenv
networkx           
numpy
transformers
torch
faiss-cpu
sentence-transformers
langchain_community
tiktoken
langchain-openai
langchainhub
chromadb
langchain
huggingface-hub
pwntools